ConstraintTrees.jl

ConstraintTrees.ConstraintTrees — Module

ConstraintTrees.jl – Tidy constraints for optimization problems

Build status	Documentation

Package ConstraintTrees.jl provides a simple data structure ConstraintTree for organizing the contents of linear and quadratic constrained optimization problems. As a main goal, it abstracts over the distinction between constraints and variables, allowing much tidier representation for many kinds of complex constraint systems. The primary purpose is to support constraint-based metabolic modeling within COBREXA.jl.

ConstraintTrees are intended to be used with JuMP, but the package does not depend on JuMP – instead it is completely generic and lightweight, and may be used with any other constraint-solving framework. The documentation describes a typical use of ConstraintTrees for solving constrained linear systems (in metabolic modeling) and constrained quadratic systems (for geometry) using JuMP, together with copy-pasteable code snippets that provide the integration.

ConstraintTrees package is maintained and open for extensions. Feel free to discuss changes and ideas via issues and pull requests.

Acknowledgements

ConstraintTrees.jl was developed at the Luxembourg Centre for Systems Biomedicine of the University of Luxembourg (uni.lu/lcsb) and at Institute for Quantitative and Theoretical Biology at Heinrich Heine University Düsseldorf (qtb.hhu.de). The development was supported by European Union's Horizon 2020 Programme under PerMedCoE project (permedcoe.eu), agreement no. 951773.

<img src="docs/src/assets/unilu.svg" alt="Uni.lu logo" height="64px"> <img src="docs/src/assets/lcsb.svg" alt="LCSB logo" height="64px"> <img src="docs/src/assets/hhu.svg" alt="HHU logo" height="64px" style="height:64px; width:auto"> <img src="docs/src/assets/qtb.svg" alt="QTB logo" height="64px" style="height:64px; width:auto"> <img src="docs/src/assets/permedcoe.svg" alt="PerMedCoE logo" height="64px">