Reference

Values

Linear and affine values

ConstraintTrees.Value — Type

struct Value

A representation of a "value" in a linear constrained optimization problem. The value is an affine linear combination of several variables.

Values can be combined additively and multiplied by real-number constants.

Multiplying two Values yields a quadratic form (in a QValue).

Fields

idxs::Vector{Int64}: Indexes of the variables used by the value. The indexes must always be sorted in strictly increasing order. The affine element has index 0.

weights::Vector{Float64}: Coefficients of the variables selected by idxs.

ConstraintTrees.Value — Method

Value(x::Real) -> ConstraintTrees.Value

Construct a constant Value with a single affine element.

ConstraintTrees.Value — Method

Value(
    x::SparseArrays.SparseVector{Float64}
) -> ConstraintTrees.Value

Shortcut for making a Value out of a linear combination defined by the SparseVector.

ConstraintTrees.value_product — Method

value_product(x::ConstraintTrees.Value, y) -> Any

Shortcut for making a dot-product between a value and anything indexable by the value indexes.

Quadratic values

ConstraintTrees.QValue — Type

struct QValue

A representation of a quadratic form in the constrained optimization problem. The QValue is an affine quadratic combination (i.e., a polynomial of maximum degree 2) over the variables.

QValues can be combined additively and multiplied by real-number constants. The cleanest way to construct a QValue is to multiply two Values.

Fields

idxs::Vector{Tuple{Int64, Int64}}: Indexes of variable pairs used by the value. The indexes must always be sorted in strictly co-lexicographically increasing order, and the second index must always be greater than or equal to the first one. (Speaking in matrix terms, the indexing follows the indexes in an upper triangular matrix by columns.)

As an outcome, the second index of the last index pair can be used as the upper bound of all variable indexes.

As with Value, index 0 represents the affine element.

weights::Vector{Float64}: Coefficient of the variable pairs selected by idxs.

ConstraintTrees.QValue — Method

QValue(x::ConstraintTrees.Value) -> ConstraintTrees.QValue

Construct a QValue that is equivalent to a given Value.

ConstraintTrees.QValue — Method

QValue(x::Real) -> ConstraintTrees.QValue

Construct a constant QValue with a single affine element.

ConstraintTrees.QValue — Method

QValue(
    x::SparseArrays.SparseMatrixCSC{Float64}
) -> ConstraintTrees.QValue

Shortcut for making a QValue out of a square sparse matrix. The matrix is force-symmetrized by calculating x' + x.

ConstraintTrees.colex_le — Method

colex_le(, ) -> Any

Internal helper for co-lex ordering of indexes.

ConstraintTrees.qvalue_product — Method

qvalue_product(x::ConstraintTrees.QValue, y) -> Any

Shortcut for computing a product of the QValue and anything vector-like.

Constraints

ConstraintTrees.Bound — Type

Shortcut for possible bounds: either no bound is present (nothing), or a single number is interpreted as an exact equality bound, or a tuple of 2 numbers is interpreted as an interval bound.

ConstraintTrees.IntervalBound — Type

struct Tuple{Float64, Float64}

Convenience shortcut for "interval" bound; consisting of lower and upper bound value.

Linear and affine constraints

ConstraintTrees.Constraint — Type

struct Constraint

A representation of a single constraint that limits the Value by a specific Bound.

Constraints may be scaled linearly, i.e., multiplied by real-number constants.

Constraints without a bound (nothing in the bound field) are possible; these have no impact on the optimization problem but the associated value becomes easily accessible for inspection and building other constraints.

Fields

value::ConstraintTrees.Value: A Value that describes what the constraint constraints.

bound::Union{Nothing, Float64, Tuple{Float64, Float64}}: A bound that the value must satisfy.

ConstraintTrees.bound — Method

bound(
    x::ConstraintTrees.Constraint
) -> Union{Nothing, Float64, Tuple{Float64, Float64}}

Simple accessor for getting out the bound from the constraint that can be used for broadcasting (as opposed to the dot-field access).

ConstraintTrees.value — Method

value(
    x::ConstraintTrees.Constraint
) -> ConstraintTrees.Value

Simple accessor for getting out the value from the constraint that can be used for broadcasting (as opposed to the dot-field access).

Quadratic constraints

ConstraintTrees.Constraint — Type

Constraint(
    v::ConstraintTrees.QValue
) -> ConstraintTrees.QConstraint
Constraint(
    v::ConstraintTrees.QValue,
    b::Union{Nothing, Float64, Tuple{Float64, Float64}}
) -> ConstraintTrees.QConstraint

Overloaded constructor of Constraint that actually makes a QConstraint because that is implied by the type of the value in v.

ConstraintTrees.QConstraint — Type

struct QConstraint

A representation of a single quadratic constraint that limits the QValue by a specific Bound. Apart from the quadratic nature, type behaves just like the normal Constraint.

Fields

qvalue::ConstraintTrees.QValue: A QValue that describes what the constraint constraints.

bound::Union{Nothing, Float64, Tuple{Float64, Float64}}: A bound that the qvalue must satisfy.

Constraint trees

ConstraintTrees.ConstraintTreeElem — Type

A shortcut for elements of the ConstraintTree.

ConstraintTrees.ConstraintTree — Type

struct ConstraintTree

A hierarchical tree of many constraints that together describe a constrained system. The tree may recursively contain other trees in a directory-like structure; and these contain Constraints and QConstraints.

Members of the constraint tree are accessible via the record dot syntax as properties; e.g. a constraint labeled with :abc in a constraint tree t may be accessed as t.abc and as t[:abc], and can be found while iterating through elems(t).

Constructing the constraint trees

Use operator ^ to put a name on a constraint to convert it into a single element ConstraintTree:

x = :my_constraint ^ Constraint(Value(...), 1.0)
dir = :my_constraint_dir ^ x

dir.my_constraint_dir.my_constraint.bound   # returns 1.0

Use operator * to glue two constraint trees together while sharing the variable indexes specified by the contained Values and QValues.

my_constraints = :linear_limit ^ Constraint(...) * :quadratic_limit ^ QConstraint(...)

Use operator + to glue two constraint trees together without sharing of any variables. The operation will renumber the variables in the trees so that the sets of variable indexes used by either tree are completely disjunct, and then glue the trees together as with *:

two_independent_systems = my_system + other_system

Variable sharing limitations

Because of the renumbering, you can not easily use constraints and values from the values before the addition in the constraint tree that is the result of the addition. There is no check against that – the resulting ConstraintTree will be valid, but will probably describe a different optimization problem than you intended.

As a rule of thumb, avoid necessary parentheses in expressions that work with the constraint trees: While t1 * t2 + t3 might work just as intended, t1 * (t2 + t3) is almost certainly wrong because the variables in t1 that are supposed to connect to variables in either of t2 and t3 will not connect properly because of renumbering of both t2 and t3. If you need to construct a tree like that, do the addition first, and construct the t1 after that, based on the result of the addition.

Fields

elems::DataStructures.SortedDict{Symbol, Union{ConstraintTrees.Constraint, ConstraintTrees.ConstraintTree, ConstraintTrees.QConstraint}}: Sorted dictionary of elements of the constraint tree.

ConstraintTrees.elems — Method

elems(
    x::ConstraintTrees.ConstraintTree
) -> DataStructures.SortedDict{Symbol, Union{ConstraintTrees.Constraint, ConstraintTrees.ConstraintTree, ConstraintTrees.QConstraint}}

Get the elements dictionary out of the ConstraintTree. This is useful for getting an iterable container for working with many constraints at once.

Also, because of the overload of getproperty for ConstraintTree, this serves as a simpler way to get the elements without an explicit use of getfield.

ConstraintTrees.incr_var_idx — Method

incr_var_idx(x::Int64, incr::Int64) -> Int64

Internal helper for manipulating variable indices.

ConstraintTrees.incr_var_idxs — Method

incr_var_idxs(
    x::ConstraintTrees.Constraint,
    incr::Int64
) -> ConstraintTrees.Constraint

Offset all variable indexes in a Constraint by the given increment.

ConstraintTrees.incr_var_idxs — Method

incr_var_idxs(
    x::ConstraintTrees.ConstraintTree,
    incr::Int64
) -> ConstraintTrees.ConstraintTree

Offset all variable indexes in a ConstraintTree by the given increment.

ConstraintTrees.incr_var_idxs — Method

incr_var_idxs(
    x::ConstraintTrees.QConstraint,
    incr::Int64
) -> ConstraintTrees.QConstraint

Offset all variable indexes in a QConstraint by the given increment.

ConstraintTrees.var_count — Method

var_count(x::ConstraintTrees.ConstraintTree) -> Any

Find the expected count of variables in a ConstraintTree.

ConstraintTrees.var_count — Method

var_count(x::ConstraintTrees.Constraint) -> Int64

Find the expected count of variables in a Constraint.

(This is a O(1) operation, relying on the order of indexes in Values.)

ConstraintTrees.var_count — Method

var_count(x::ConstraintTrees.QConstraint) -> Int64

Find the expected count of variables in a Constraint.

(This is a O(1) operation, relying on the co-lexicographical ordering of indexes in QValues)

ConstraintTrees.variable — Method

variable(; bound) -> ConstraintTrees.Constraint

Allocate a single unnamed variable, returning a Constraint with an optionally specified bound.

ConstraintTrees.variables — Method

variables(; keys, bounds)

Make a trivial constraint system that creates variables with indexes in range 1:length(keys) named in order as given by keys.

Parameter bounds is either nothing for creating unconstrained variables, a single bound (of precise length 1) for creating all variables of the same constraint, or an iterable object of same length as keys with individual bounds for each variable in the same order as keys.

The individual bounds should be of type Bound. To pass a single interval bound for all variables, it is impossible to use a tuple (since its length is 2); in such case use bound = Ref((minimum, maximum)), which has the correct length.

Solution trees

ConstraintTrees.SolutionTreeElem — Type

A shortcut for type of contents in a SolutionTree.

ConstraintTrees.SolutionTree — Type

struct SolutionTree

A structure similar to ConstraintTree, but only holds the resolved values of each constraint. As with ConstraintTree, use record dot notation and elems to browse the solution structure.

Use SolutionTree to construct the SolutionTree out of a ConstraintTree and a vector of variable values.

To construct a SolutionTree, combine a ConstraintTree (or generally any ConstraintTreeElem with a vector of variable assignments (typically representing a constrained problem solution) using the overloaded 2-parameterSolutionTree constructor. The result will contain a tree of constraint values w.r.t. the given variable assignment (or just a single number in case the input was only a single constraint).

Example

cs = ConstraintTree(...)
vals = [1.0, 2.0, 4.0]
SolutionTree(cs, vals)

Fields

elems::DataStructures.SortedDict{Symbol, Union{Float64, ConstraintTrees.SolutionTree}}

ConstraintTrees.elems — Method

elems(
    x::ConstraintTrees.SolutionTree
) -> DataStructures.SortedDict{Symbol, Union{Float64, ConstraintTrees.SolutionTree}}

Get the elements dictionary out of the SolutionTree.

The use is similar as with the overload for ConstraintTree.