Reference

Values (linear combinations of variables)

ConstraintTrees.Value — Type

struct Value

A representation of a "value" in a linear constrained optimization problem. The value is a linear combination of several variables.

Values can be combined additively and multiplied by real-number constants.

Fields

idxs::Vector{Int64}: Indexes of the variables used by the value. The indexes are always sorted in increasing order.

weights::Vector{Float64}: Coefficients of the variables as used by the value

ConstraintTrees.Value — Method

Value(
    x::SparseArrays.SparseVector{Float64}
) -> ConstraintTrees.Value

Shortcut for making a Value out of SparseVector.

ConstraintTrees.value_product — Method

value_product(x::ConstraintTrees.Value, y) -> Any

Shortcut for making a dot-product between a value and anything indexable by the value indexes.

Constraints

ConstraintTrees.Bound — Type

Shortcut for possible bounds: either no bound is present (nothing), or a single number is interpreted as an exact equality bound, or a tuple of 2 values is interpreted as an interval bound.

ConstraintTrees.Constraint — Type

struct Constraint

A representation of a single constraint that limits the Value by a specific Bound.

Constraints may be multiplied by real-number constants.

Constraints without a bound (nothing in the bound field) are possible; these have no impact on the optimization problem but the associated value becomes easily accessible for inspection and building other constraints.

Fields

value::ConstraintTrees.Value: A Value that describes what the constraint constraints.

bound::Union{Nothing, Float64, Tuple{Float64, Float64}}: A bound that the value must satisfy.

ConstraintTrees.IntervalBound — Type

struct Tuple{Float64, Float64}

Convenience shortcut for "interval" bound; consisting of lower and upper bound value.

ConstraintTrees.value — Method

value(
    x::ConstraintTrees.Constraint
) -> ConstraintTrees.Value

Simple accessor for getting out the value from the constraint that can be used for broadcasting (as opposed to the dot-field access).

Constraint trees

ConstraintTrees.ConstraintTreeElem — Type

A shortcut for elements of the ConstraintTree.

ConstraintTrees.ConstraintTree — Type

struct ConstraintTree

A hierarchical tree of many constraints that together describe a constrained linear system. The tree may recursively contain other trees in a directory-like structure.

Members of the constraint tree are accessible via the record dot syntax as properties; e.g. a constraint labeled with :abc in a constraint tree t may be accessed as t.abc and as t[:abc], and can be found while iterating through elems(t).

Constructing the constraint trees

Use operator ^ to put a name on a Constraint to convert it into a single element ConstraintTree:

x = :my_constraint ^ Constraint(Value(...), 1.0)
dir = :my_constraint_dir ^ x

dir.my_constraint_dir.my_constraint.bound   # returns 1.0

Use operator * to glue two constraint trees together while sharing the variable indexes specified by the contained Values.

my_constraints = :constraint1 ^ Constraint(...) * :constraint2 ^ Constraint(...)

Use operator + to glue two constraint trees together without sharing of any variables. The operation will renumber the variables in the trees so that the sets of variable indexes used by either tree are completely disjunct, and then glue the trees together as with *:

two_independent_systems = my_system + other_system

Because of the renumbering, you can not easily use constraints and values from the values before the addition in the constraint tree that is the result of the addition. There is no check against that; the resulting ConstraintTree will be valid, but will probably describe a different optimization problem than you intended.

As a rule of thumb, avoid necessary parentheses in expressions that work with the constraint trees: While t1 * t2 + t3 might work just as intended, t1 * (t2 + t3) is almost certainly wrong because the variables in t1 that are supposed to connect to variables in either of t2 and t3 will not connect properly because of renumbering of both t2 and t3. If you need to construct a tree like that, do the addition first, and construct the t1 after that, based on the result of the addition.

Fields

elems::DataStructures.SortedDict{Symbol, Union{ConstraintTrees.Constraint, ConstraintTrees.ConstraintTree}}: Sorted dictionary of elements of the constraint tree.

ConstraintTrees.allocate_variables — Method

allocate_variables(; keys, bounds)

Make a trivial constraint system that creates variables with indexes in range 1:length(keys) named in order as given by keys.

Parameter bounds is either nothing for creating unconstrained variables, a single bound (of precise length 1) for creating all variables of the same constraint, or an iterable object of same length as keys with individual bounds for each variable in the same order as keys.

The individual bounds should be of type Bound. To pass a single interval bound for all variables, it is impossible to use a tuple (since its length is 2); in such case use bound = Ref((minimum, maximum)), which has the correct length.

ConstraintTrees.elems — Method

elems(
    x::ConstraintTrees.ConstraintTree
) -> DataStructures.SortedDict{Symbol, Union{ConstraintTrees.Constraint, ConstraintTrees.ConstraintTree}}

Get the elements dictionary out of the ConstraintTree. This is useful for getting an iterable container for working with many constraints at once.

Also, because of the overload of getproperty for ConstraintTree, this serves as a simpler way to get the elements without an explicit use of getfield.

ConstraintTrees.make_constraint_tree — Method

make_constraint_tree(x...) -> ConstraintTrees.ConstraintTree

Create a properly typed ConstraintTree out of anything that can be used to construct the inner dictionary.

Example

make_constraint_tree(:a => some_constraint, :b => another_constraint)
make_constraint_tree(c for c=constraints if !isnothing(c.bound))

Solution trees

ConstraintTrees.SolutionTreeElem — Type

A shortcut for type of contents in a SolutionTree.

ConstraintTrees.SolutionTree — Type

struct SolutionTree

A structure similar to ConstraintTree, but only holds the resolved values of each constraint. As with ConstraintTree, use record dot notation and elems to browse the solution structure.

Use solution_tree to construct the SolutionTree out of a ConstraintTree and a vector of variable values.

Fields

elems::DataStructures.SortedDict{Symbol, Union{Float64, ConstraintTrees.SolutionTree}}

ConstraintTrees.elems — Method

elems(
    x::ConstraintTrees.SolutionTree
) -> DataStructures.SortedDict{Symbol, Union{Float64, ConstraintTrees.SolutionTree}}

Get the elements dictionary out of the SolutionTree.

The use is similar as with the overload for ConstraintTree.

ConstraintTrees.solution_tree — Method

solution_tree(
    x::ConstraintTrees.Constraint,
    vars::AbstractVector{Float64}
) -> Any

Convert a single constraint and a vector of variable assignments (typically representing a constrained problem solution) to the value of the constraint w.r.t. the given variable assignment.

ConstraintTrees.solution_tree — Method

solution_tree(
    x::ConstraintTrees.ConstraintTree,
    vars::AbstractVector{Float64}
) -> ConstraintTrees.SolutionTree

Convert a ConstraintTree and a vector of variable assignments (typically representing a constrained problem solution) to a SolutionTree of constraint values w.r.t. the given variable assignment.